+++ frontend/src/metabase/dashboard/components/DashboardHeader/buttons/DashboardActionMenu.tsx
@@ -1,23 +1,35 @@
 import type { MouseEvent } from "react";
 import { t } from "ttag";
 
-import EntityMenu from "metabase/components/EntityMenu";
+import EntityMenuTrigger from "metabase/components/EntityMenuTrigger";
 import { trackExportDashboardToPDF } from "metabase/dashboard/analytics";
+import {
+  type OverflowMenuItem,
+  OverflowMenu,
+} from "metabase/dashboard/components/DashboardHeader/buttons/OverflowMenu";
 import { DASHBOARD_PDF_EXPORT_ROOT_ID } from "metabase/dashboard/constants";
 import type { DashboardFullscreenControls } from "metabase/dashboard/types";
 import { PLUGIN_DASHBOARD_HEADER } from "metabase/plugins";
+import { Box } from "metabase/ui";
 import { saveDashboardPdf } from "metabase/visualizations/lib/save-dashboard-pdf";
 import type { Dashboard } from "metabase-types/api";
 
-export const DashboardActionMenu = (props: { items: any[] }) => (
-  <EntityMenu
-    key="dashboard-action-menu-button"
-    triggerAriaLabel="dashboard-menu-button"
-    items={props.items}
-    triggerIcon="ellipsis"
-    tooltip={t`Move, trash, and more...`}
-    // TODO: Try to restore this transition once we upgrade to React 18 and can prioritize this update
-    transitionDuration={0}
+type DashboardActionMenuProps = {
+  items: OverflowMenuItem[];
+};
+
+export const DashboardActionMenu = ({ items }: DashboardActionMenuProps) => (
+  <OverflowMenu
+    items={items}
+    target={
+      <Box>
+        <EntityMenuTrigger
+          ariaLabel={t`Move, trash, and more...`}
+          icon="ellipsis"
+          tooltip={t`Move, trash, and more...`}
+        />
+      </Box>
+    }
   />
 );
 
@@ -31,17 +43,14 @@ export const getExtraButtons = ({
   dashboard: Dashboard;
   canEdit: boolean;
   pathname: string;
-}) => {
-  const extraButtons = [];
-
-  extraButtons.push({
+}): OverflowMenuItem[] => [
+  {
     title: t`Enter fullscreen`,
     icon: "expand",
     action: (e: MouseEvent) => onFullscreenChange(!isFullscreen, !e.altKey),
     event: `Dashboard;Fullscreen Mode;${!isFullscreen}`,
-  });
-
-  extraButtons.push({
+  },
+  {
     title:
       Array.isArray(dashboard.tabs) && dashboard.tabs.length > 1
         ? t`Export tab as PDF`
@@ -54,34 +63,29 @@ export const getExtraButtons = ({
         trackExportDashboardToPDF(dashboard.id);
       });
     },
-  });
-
-  if (canEdit) {
-    extraButtons.push({
-      title: t`Move`,
-      icon: "move",
-      link: `${pathname}/move`,
-      event: "Dashboard;Move",
-    });
-  }
-
-  extraButtons.push({
+  },
+  {
+    enabled: canEdit,
+    title: t`Move`,
+    icon: "move",
+    link: `${pathname}/move`,
+    event: "Dashboard;Move",
+  },
+  {
     title: t`Duplicate`,
     icon: "clone",
     link: `${pathname}/copy`,
     event: "Dashboard;Copy",
-  });
-
-  if (canEdit) {
-    extraButtons.push(...PLUGIN_DASHBOARD_HEADER.extraButtons(dashboard));
-
-    extraButtons.push({
-      title: t`Move to trash`,
-      icon: "trash",
-      link: `${pathname}/archive`,
-      event: "Dashboard;Archive",
-    });
-  }
-
-  return extraButtons;
-};
+  },
+  {
+    enabled: canEdit,
+    ...PLUGIN_DASHBOARD_HEADER.extraButtons(dashboard),
+  },
+  {
+    enabled: canEdit,
+    title: t`Move to trash`,
+    icon: "trash",
+    link: `${pathname}/archive`,
+    event: "Dashboard;Archive",
+  },
+];
